docs: finish the shell guide
I left out bash scripting the guide was already getting quite long and I
find the use case for the students to be smaller in this regard. I have
left some good ressources for them to read instead if they want to delve
into that world. Finished the makefile chapter with a very short example